requirements.txt
Django==1.6.10
South==1.0.2
coverage==3.7.1
djangorestframework==3.0.5
wsgiref==0.1.2
